refactor: 장바구니 아이템을 종류별로 카운팅하는 함수를 utils에서 domain으로 이동
- 범용적인 유틸 함수 보다는 도메인 로직에 가깝다고 느껴 이동함
