--- src/util/countDistinct.ts
@@ -1,3 +1,0 @@
-export const countDistinct = (cartItemsIds: number[]) => {
-  return new Set(cartItemsIds).size;
-};
